+++ app/javascript/vue/components/radials/annotator/components/alternate_value_annotator.vue
@@ -1,30 +1,21 @@
 <template>
   <div class="alternate_values_annotator">
     <div>
-      <div class="switch-radio">
-        <template v-for="(item, key, index) in typeList">
-          <input
-            v-model="alternate_value.type"
-            :value="key"
-            :id="`alternate_values-picker-${index}`"
-            name="alternate_values-picker-options"
-            type="radio"
-            class="normal-input button-active"
-          >
-          <label
-            :for="`alternate_values-picker-${index}`"
-            class="capitalize">{{ item }}
-          </label>
-        </template>
-      </div>
+      <switch-component
+        :options="tabs"
+        use-index
+        v-model="tabIndex"
+      />
     </div>
-    <div v-if="alternate_value.type">
+    <div v-if="alternateType">
       <ul class="no_bullets content">
-        <li v-for="(item, key) in values">
+        <li
+          v-for="(item, key) in values"
+          :key="item">
           <label>
             <input
               type="radio"
-              v-model="alternate_value.alternate_value_object_attribute"
+              v-model="alternateValue.alternate_value_object_attribute"
               :value="key">
             "{{ key }}" -> {{ item }}
           </label>
@@ -32,21 +23,22 @@
       </ul>
 
       <autocomplete
+        v-if="alternateType == ALTERNATE_VALUE_TRANSLATION"
         class="field"
-        v-if="alternate_value.type == 'AlternateValue::Translation'"
         url="/languages/autocomplete"
         label="label"
         min="2"
         placeholder="Language"
-        @getItem="alternate_value.language_id = $event.id"
-        param="term"/>
+        @getItem="alternateValue.language_id = $event.id"
+        param="term"
+      />
 
       <div class="separate-bottom">
         <div class="field">
           <input
             type="text"
             class="normal-input"
-            v-model="alternate_value.value"
+            v-model="alternateValue.value"
             placeholder="Value">
         </div>
         <button
@@ -62,8 +54,6 @@
     <display-list
       label="object_tag"
       :list="list"
-      :edit="true"
-      @edit="note = $event"
       @delete="removeItem"
       class="list"/>
   </div>
@@ -72,25 +62,42 @@
 
 import CRUD from '../request/crud.js'
 import annotatorExtend from '../components/annotatorExtend.js'
-import autocomplete from 'components/ui/Autocomplete.vue'
-import displayList from './displayList.vue'
+import Autocomplete from 'components/ui/Autocomplete.vue'
+import SwitchComponent from 'components/switch.vue'
+import DisplayList from './displayList.vue'
+import {
+  ALTERNATE_VALUE_ABBREVIATION,
+  ALTERNATE_VALUE_ALTERNATE_SPELLING,
+  ALTERNATE_VALUE_MISSPELLING,
+  ALTERNATE_VALUE_TRANSLATION
+} from 'constants/index.js'
+import { AlternateValue } from 'routes/endpoints'
 
 export default {
   mixins: [CRUD, annotatorExtend],
 
   components: {
-    displayList,
-    autocomplete
+    Autocomplete,
+    DisplayList,
+    SwitchComponent
   },
 
   computed: {
-    validateFields() {
-      return (this.alternate_value.value &&
-        this.alternate_value.alternate_value_object_attribute)
+    validateFields () {
+      return this.alternateValue.value &&
+        this.alternateValue.alternate_value_object_attribute
+    },
+
+    tabs () {
+      return Object.values(this.typeList)
+    },
+
+    alternateType () {
+      return Object.keys(this.typeList)[this.tabIndex]
     }
   },
 
-  mounted () {
+  created () {
     this.getList(`/alternate_values/${encodeURIComponent(this.globalId)}/metadata`).then(response => {
       this.values = response.body
     })
@@ -100,56 +107,38 @@ export default {
     return {
       values: undefined,
       typeList: {
-        'AlternateValue::Translation': 'translation',
-        'AlternateValue::Abbreviation': 'abbreviation',
-        'AlternateValue::Misspelling': 'misspelled',
-        'AlternateValue::AlternateSpelling': 'alternate spelling'
+        [ALTERNATE_VALUE_TRANSLATION]: 'translation',
+        [ALTERNATE_VALUE_ABBREVIATION]: 'abbreviation',
+        [ALTERNATE_VALUE_MISSPELLING]: 'misspelled',
+        [ALTERNATE_VALUE_ALTERNATE_SPELLING]: 'alternate spelling'
       },
-      alternate_value: this.newAlternate()
+      ALTERNATE_VALUE_TRANSLATION,
+      alternateValue: this.newAlternate(),
+      tabIndex: 0
     }
   },
 
   methods: {
     newAlternate () {
       return {
-        type: undefined,
         value: undefined,
         language_id: undefined,
-        alternate_value_object_attribute: undefined,
-        annotated_global_entity: decodeURIComponent(this.globalId)
+        alternate_value_object_attribute: undefined
       }
     },
 
     createNew () {
-      this.create('/alternate_values', {alternate_value: this.alternate_value}).then(response => {
+      const alternate_value = {
+        ...this.alternateValue,
+        type: this.alternateType,
+        annotated_global_entity: decodeURIComponent(this.globalId)
+      }
+
+      AlternateValue.create({ alternate_value }).then(response => {
         this.list.push(response.body)
-        this.alternate_value = this.newAlternate()
+        this.alternateValue = this.newAlternate()
       })
     }
   }
 }
 </script>
-<style lang="scss">
-  .radial-annotator {
-    .alternate_values_annotator {
-      .field input {
-        width: 100%;
-      }
-      .switch-radio {
-        label {
-          min-width: 95px;
-        }
-      }
-      li {
-        border-right: 0px;
-        padding-left: 0px;
-      }
-      textarea {
-        padding-top: 14px;
-        padding-bottom: 14px;
-        width: 100%;
-        height: 100px;
-      }
-    }
-  }
-</style>